feat(exchange-promo): move promo to send folder in components
