+++ packages/blockchain-wallet-v4-frontend/src/modals/Bch/SendBch/FirstStep/template.success.js
@@ -31,7 +31,7 @@ import { model } from 'data'
 import { required, validBchAddress } from 'services/FormHelper'
 import Bowser from 'bowser'
 import ComboDisplay from 'components/Display/ComboDisplay'
-import ExchangePromo from 'components/ExchangePromo'
+import ExchangePromo from 'components/Send/ExchangePromo'
 import MnemonicRequiredForCustodySend from 'components/Send/RecoveryPhrase'
 import PropTypes from 'prop-types'
 import QRCodeCapture from 'components/QRCodeCapture'
