+++ packages/blockchain-wallet-v4-frontend/src/modals/Eth/SendEth/FirstStep/template.success.js
@@ -46,7 +46,7 @@ import { Remote } from 'blockchain-wallet-v4/src'
 import { required, validEthAddress } from 'services/FormHelper'
 import Bowser from 'bowser'
 import ComboDisplay from 'components/Display/ComboDisplay'
-import ExchangePromo from 'components/ExchangePromo'
+import ExchangePromo from 'components/Send/ExchangePromo'
 import LowBalanceWarning from './LowBalanceWarning'
 import LowEthWarningForErc20 from './LowEthWarningForErc20'
 import MnemonicRequiredForCustodySend from 'components/Send/RecoveryPhrase'
